Create Birdcatcher.py
This program checks to see if a specified alias is registered on twitter if it is it will return a link to the profile. If the account does not exist it will reply with "[!] Page does not exist", and if the account is suspended it will reply "[!] Account suspended [!]" along with a link to the account.